feat(briefing): extract events pipeline (commit 8/11)
Workstream 6 Step 1 — extracts the events pipeline (THE BIG ONE) from
briefing-service.js + deletes 4 INDEPENDENTLY dead code paths along the way
(Option A precedent — see claude_memory #294, #295, #296).

DRIFT RESOLUTION: Recon-first methodology surfaced 4 dead paths, not 1:
  1. fetchEventsWithClaudeWebSearch (~123 lines) — Claude WebSearch fallback
     never wired into fetchEventsForBriefing
  2. _fetchEventsWithGemini3ProPreviewLegacy (~131 lines) — superseded by
     parallel-category fetchEventsWithGemini3ProPreview
  3. mapGeminiEventsToLocalEvents (~53 lines) — orphan event-shape mapper
  4. LocalEventSchema (~18 lines) — orphan Zod schema, paired with mapGemini
All 4 verified DEAD via 4-way cross-check (internal grep, external grep,
string-quoted refs, dynamic-import enumeration). All 4 are independent
(none calls another).

METHODOLOGY LESSON (logged claude_memory #297): naive `\b`-bounded grep
produced a false-negative on fetchEventsWithGemini3ProPreview (the LIVE
primary fetcher) because it's a substring of _fetchEventsWithGemini3ProPreviewLegacy.
Substring-collision is a real grep failure mode when one function name
contains another. Resolution: re-grep without `\b`, READ function bodies,
cross-check with quote/dynamic dispatch. Master Architect's red flag
("that name sounds like the primary live Gemini fetcher") was correct.

NEW: server/lib/briefing/pipelines/events.js (~1,090 lines)
  - discoverEvents({ snapshot, snapshotId }) — pipeline contract entry,
    returns { events: {items, reason} | errorMarker, reason }. Owns the
    polymorphic SSE-write (array when items > 0, {items, reason} object
    when empty) and the errorMarker .catch wrapper.
  - fetchEventsForBriefing — re-exported live path. Full discovery + DB
    caching + read pipeline: deactivatePastEvents → fetchEventsWithGemini3ProPreview
    → validateEventsHard → deduplicateEvents (HASH) → deduplicateEventsSemantic
    → venue resolution chain → discovered_events INSERT → JOIN read →
    filterInvalidEvents → return.
  - deduplicateEvents — re-exported HASH dedup (Rule 16, SURVIVAL GUARDRAIL #1,
    distinct from deduplicateEventsSemantic.js).
  - filterInvalidEvents — re-exported LIVE compatibility shim (SURVIVAL
    GUARDRAIL #2, distinct from validateEventsHard).
  - fetchEventsWithGemini3ProPreview — module-private (single live caller
    is fetchEventsForBriefing, same shape as Commit 5's extractAirportJson
    private-helper precedent).
  - fetchEventCategory — module-private (single live caller is
    fetchEventsWithGemini3ProPreview).
  - withTimeout, EVENT_SEARCH_TIMEOUT_MS, EVENT_CATEGORIES — module-private
    helpers, moved with the events code (no remaining callers in
    briefing-service.js after deletion of dead Claude block).

briefing-service.js delta (1,936 → 802 lines, net -1,134):
  - import { discoverEvents, fetchEventsForBriefing, deduplicateEvents,
    filterInvalidEvents } from pipelines/events.js
  - export { fetchEventsForBriefing, deduplicateEvents, filterInvalidEvents }
    (re-export for external + internal callers)
  - orchestrator eventsPromise: replaced .then/.catch chain with direct
    discoverEvents() call (~10-line shrink, pipeline owns its own SSE write
    and errorMarker)
  - eventsItems extraction: eventsResult?.items → eventsResult?.events?.items
    (line 435, warning message updated to match)
  - final-assembly: eventsResult?.reason → eventsResult?.events?.reason
    (line 505, reads inner section reason)
  - refreshEventsInBriefing untouched: uses fetchEventsForBriefing directly
    via the re-export binding; flat {items, reason, provider} shape unchanged
  - Sed deletes (3 ranges, BOTTOM-first): 477-1272 (events tail), 345-475
    (dead Claude block), 125-343 (timeout helpers + dedup + filter).
    Total ~1,146 lines removed.

CUMULATIVE COMMITS 4-8: 3,683 → 802 lines (-2,881 net, ~78% reduction).

SURVIVAL GUARDRAILS preserved per Master Architect directive:
  - deduplicateEvents (HASH dedup, Rule 16) — distinct from
    deduplicateEventsSemantic.js (semantic title-similarity). Both coexist;
    both run in sequence inside fetchEventsForBriefing (hash first then
    semantic). Re-exported.
  - filterInvalidEvents (LIVE compat shim) — distinct from validateEventsHard
    (canonical validation module). Re-exported because external API callers
    (briefing.js, dump-last-briefing.js) still use the legacy name.

Verification:
  - npm run typecheck (tsc --noEmit): clean
  - npm run lint (eslint client/src): clean
  - node --check on briefing-service.js, pipelines/events.js: pass
  - Dynamic import smoke: events.js exports
    { deduplicateEvents, discoverEvents, fetchEventsForBriefing, filterInvalidEvents }
  - Re-export chain verified callable: deduplicateEvents=function,
    filterInvalidEvents=function, fetchEventsForBriefing=function
  - Dynamic-import enumeration: fetchEventsWithGemini3ProPreview +
    fetchEventCategory confirmed PRIVATE (zero export, zero dynamic refs)
  - Grep audits: zero orphan local definitions, zero literal channel strings
    outside briefing-notify.js, zero references to deleted symbols outside
    intentional documentation comments
  - Both error pathways preserved:
    Pathway A (thrown) → Promise.allSettled captures → eventsResult undefined
    → failedReasons.events populated → fallback at line 505 fires
    Pathway B (graceful) → discoverEvents returns
    { events: {items, reason}, reason } → eventsResult.events.items +
    eventsResult.events.reason resolved at lines 435/505

Per PLAN_workstream6_remaining-2026-05-02.md §2.1 (wider contract),
§2.5 (re-export pattern), §2.3 (sed for line-range deletes).
Same Option A precedent as commits 4 + 6, plus commit 5 + 7 private-helper
co-location pattern (fetchEventsWithGemini3ProPreview + fetchEventCategory
move with their only live caller).
